+++ cmd/main/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/Sanchir01/sandjma_graphql/internal/config"
 	categoryStore "github.com/Sanchir01/sandjma_graphql/internal/database/store/category"
 	colorStorage "github.com/Sanchir01/sandjma_graphql/internal/database/store/color"
-	"github.com/Sanchir01/sandjma_graphql/internal/server/grpc/sandjmagrpc"
+	"github.com/Sanchir01/sandjma_graphql/internal/server/grpc/authgrpc"
 	"github.com/Sanchir01/sandjma_graphql/pkg/lib/db/connect"
 
 	"github.com/Sanchir01/sandjma_graphql/internal/database/store/product"
@@ -46,7 +46,8 @@ func main() {
 	defer db.Close()
 	trManager := manager.Must(trmsqlx.NewDefaultFactory(db))
 	r := chi.NewRouter()
-	authgrpclient, err := sandjmagrpc.NewGrpcAuth(context.Background(), cfg.GrpcClients.Auth.Address, cfg.GrpcClients.Auth.Retries, lg)
+
+	authgrpclient, err := authgrpc.NewGrpcAuth(context.Background(), cfg.GrpcClients.Auth.Address, cfg.GrpcClients.Auth.Retries, lg)
 	if err != nil {
 		lg.Error("failed init auth grpc client", err)
 		os.Exit(1)
