+++ internal/gql/api/category/category.graphqls
@@ -1,0 +1,9 @@
+type Category implements VersionInterface{
+    id:Uuid!
+    name:String!
+    slug:String!
+    created_at: DateTime!
+    updated_at: DateTime!
+    description:String!
+    version:UInt!
+}
